--- src/main/java/Task.java
@@ -1,33 +1,0 @@
-public abstract class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public Task(String description, boolean isDone) {
-        this.description = description;
-        this.isDone = isDone;
-    }
-
-    public String getStatusIcon() {
-        return isDone ? "[✔]" : "[ ]";
-    }
-
-    abstract public void taskPrinter();
-
-    abstract public void taskPrinter(int index);
-
-    public void markAsDone() {
-        this.isDone = true;
-    }
-
-    public void markAsUndone() {
-        this.isDone = false;
-    }
-
-    abstract public String storagePrinter();
-
-}
