+++ src/main/java/command/Command.java
@@ -1,3 +1,8 @@
+package command;
+import storage.Storage;
+import task.TaskList;
+import ui.Ui;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage);
     public abstract boolean isExit();
