+++ src/main/java/ui/Ui.java
@@ -1,3 +1,7 @@
+package ui;
+
+import task.Task;
+
 import java.util.Scanner;
 
 public class Ui {
